working on showing selected block
