+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2015-2016, SnowGremlin.
+Copyright (c) 2015-2018, SnowGremlin.
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
