+++ web/examples/craft/blockType.dart
@@ -1,4 +1,4 @@
-part of example1;
+part of craft;
 
 /// BlockType is the enums for all the voxel block types
 class BlockType {
